backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
jsonschema
pandas
pymupdf
python-docx
pytesseract
Pillow
pythainlp
pytest
fpdf2
python-multipart
